fix: changed folders names
