+++ src/App/index.jsx
@@ -5,10 +5,10 @@ import SettingsPage from '@/pages/Settings'
 import { HomePage } from '@/pages/Home'
 import Switch from 'react-router-dom/es/Switch'
 import Route from 'react-router-dom/es/Route'
-import { HeaderMain } from '@/components/A1_Header'
+import { HeaderMain } from '@/components/Header'
 import { useSelector } from 'react-redux'
 import { ThemeProvider } from 'styled-components'
-import { ErrorBoundary } from '@/components/A7_ErrorBoundary/ErrorBoundary'
+import { ErrorBoundary } from '@/components/ErrorBoundary/ErrorBoundary'
 
 
 export const Application = () => {
