+++ src/components/History/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { HistoryItem, HistoryTitle, HistoryWrapper, HistoryItemWrapper } from '@/components/A5_History/components'
+import { HistoryItem, HistoryTitle, HistoryWrapper, HistoryItemWrapper } from '@/components/History/components'
 import { useSelector } from 'react-redux'
-import { ErrorBoundary } from '@/components/A7_ErrorBoundary/ErrorBoundary'
+import { ErrorBoundary } from '@/components/ErrorBoundary/ErrorBoundary'
 
 export const History = () => {
   const historyItems = useSelector(state => state.history.historyItems)
