+++ src/components/Keypad/index.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import { Button, ButtonsWrapper } from '@/components/A4_Keypad/components'
+import { Button, ButtonsWrapper } from '@/components/Keypad/components'
 import { useDispatch, useSelector } from 'react-redux'
 import { addOperationAC, addValueAC, calcValueAC, resetLastValueAC, resetValueAC } from '@/actions'
 import { getMathHandler, separator } from '@/helpers'
